Add doc on the syntax of fonts
This just describes the two ways of specifying a font.

There are three parsers used in fontdata.cpp, but the third one
(XLoadQueryFont) also uses XLFD and only is a fallback if the
XCreateFontSet() does not work, but I don't think it is worth
mentioning that in the user documentation.