Fix doc/gendoc.py extractor to strip comments
Formerly, the doc/gendoc.py keept the comment tokens in the token tree.
This resulted in a comment being treated as the function body of a
function definition and hence some of the documentation was missed.

Also, this was not noticed because the classname2examplepath array was
not extended in test_doc.py. The present change fixes the main issue and
the test case, and even makes it a bit stricter.