Fix completion in EmptyOrInt
The '0' was listed multiple times in the completion of the
'move_monitor' command. The issue was a wrong completion in the
EmptyOrInt class.

The new test case checks that there is only one '0' in the completion
list. In particular, this test ensures that the completions of multiple
optional arguments (such as in 'move_monitor') do not overlap.