Only resend ConfigureNotify on geometry change
This fixes an issue (#1235) which consisted of hlwm hanging because of a
loop between hlwm and the reaper application (reaper.fm).

Apparently, the issue is that `XA_WM_NORMAL_HINTS` is updated by the
reaper window, and this causes the tiling layout to be re-computed,
which results in a ConfigureNotify sent to the window. This event seems
to make reaper set `XA_WM_NORMAL_HINTS` again and so the whole process
starts again.

The solution is to only send the ConfigureNotify if the geometry really
changed. I don't know how to write a test case for it, but any test case
more or less should check that this still sends enough ConfigureNotify
events as requested by ICCCM (section 4.1.5.).