Remember last reported geometry in client directly (#1240)
Instead of deciding in Decoration when to send a configure, just
remember the last configure event in every client so we know directly
whether it is worth sending a new configure event or not.

This fixes the bug #1239 where moving a client with the mouse did not
send the configure event when the mouse dragging was over. (This caused
the menus in Qt applications to be positioned wrong.). The bug was
originally introduced by #1236, thanks to @tiosgz for reporting it!

For easier testing, this remembered window geometry is exposed via the
attribute system. To simplify writing tests for it, there is a new type
'Rectangle' in the python bindings.