Force sending a configure request in some contexts
Depending on the context, it is mandatory to send back a configure
request, even if the window geometry did not change.

This fixes the issue that gnome-terminal and xfce4-terminal are blank on
start up, because they are waiting for the ConfigureNotify.

The bug was introduced by #1240 which is part of a sequence of changes
regarding issue #1235. The present change is just a quick fix, in the
long run this should be solved more elegantly.