Move 'focus_edge' and 'shift_edge' to GlobalCommands (#1252)
This moves these two command wrappers to GlobalCommands and migrates the
completion to the Completion objects. I think it would have been too
complicated to force them into the ArgParse format.

A test case already existed for both, and the completion is the same as
for 'focus' and 'shift'.

This closes #565.