Parse --skip-visible in move_index command
When migrating the 'move_index' command in a5bdb0d45 (#1227) I forgot to
parse the --skip-visible flag. This brings the flag back and extends the
test cases of 'use_index' to also test 'move_index'.

While at it, I've also added the suggestions "-1" and "+1" for the index
argument (as it already exists for use_index).

This fixes #1256, thanks to @gravndal for reporting this issue!