Make 'cycle' work for the floating layer
If the floating layer is focused, 'cycle' should traverse through the
clients in there. While at it, I've migrated the command to ArgParse and
moved it to HSTag. As usual, there are some new test cases.

Still, this only cycles through the windows explicitly marked as
floating, which is maybe odd when switching the entire tag to floating
(which should be fixed in the long run, see issue #948).