tests: Relaxe timeouts for shut down a bit
Sometimes the CI fails because the shutdown takes too long.
Thus, a slightly longer timeout might be reasonable. Moreover, the
shutdown of the HlwmBridge is now in parallel, so all processes are
sent the term signal simultaneously, so in sum, this might even speed up
the shut down phase of test cases with clients.