Remove old completion code from command.cpp (#1265)
With the completion code from all commands migrated, most of the code in
command.cpp is now obsolete. There is still a bit of the char**-based code
which needs to be reorganized until we get rid of the old command
interface entirely.

The new tests verify that the exit code of the completion of `spawn` and
`wmexec` still are correct.