Add client attribute 'float_geometry'
Add a new attribute to access and write the floating geometry of a
client. In order to keep the diff small, I kept the name of the C++
variable in the source code, even though '...geometry' is more
appropriate.

In the diff, I replaced member write-access with a preparation of the
new Rectangle and a single write. This is inevitable anyway because the
operators * and -> of Attribute_<> only return const references (as they
should). Read-only member access is no problem; here, I only had to
replace '.' with '->'.

There is not much necessity to test the attribute itself, but instead,
it serves as a tool to test other features of hlwm (e.g. the pad and the
monitor geometry).