New command 'list_clients'
For the new command 'list_clients' the main challange was to extend
ArgParse by command line flags that take arguments. To make the
completion work, I also had to extend the completion objects by a
'prepend' string that is prepended to every completion result.

This implements the main feature request of #831 and also includes the
functionality of the reject PR #1249.