Auto-convert attributes to/from python types
In the python bindings, automatically convert the attribute values from
(resp. to) their corresponding python types when writing (resp. reading)
attributes.

This requires touching (but fortunately simplifying) a lot of assertions
in the test cases. Now, working with the contents of hlwm.attr should
feel as if one had the hlwm attributes available as python variables.

A few new test cases at the end of test_python_binds.py illustrate the
usage for all types that are available at the moment.