Support error output channel internally (#1278)
In the command system, the 'Output' parameter now also carries an error
channel and the command name (in addition to the standard output
channel). This is accomplished with a new class OutputChannels in
commandio.h (and Output = OutputChannels&).

So far, this only affects the command system, but not the IPC interface,
which thus simply concatenates the two output channels in the reply to
'herbstclient'.

Thus, we can now print error messages to this dedicated channel and have
the error message prefixed with the command name automatically. Most of
the diff consists of migrating the error message printing to the new
perror(), which does the prefixing.

As a bonus, CallOrComplete does not need the command name anymore.

Also, HlwmCommon::callCommand() was only used in XMainLoop, so I have
used it there.