Extend IPC protocol by error channel (#1285)
This makes the ipc protocol and herbstclient support the error channel.
In order to make herbstclient compatible with old herbstluftwm ipc
server implementations, the error channel support is indicated by an
atom on hlwm's ipc server window (which is so far only used for the hook
system). Conversely, an old herbstclient version should also work more
or less with the new herbstluftwm ipc server: it simply ignores the
error channel.

Due to this change, a few more missing `output.perror()` calls surfaced,
affecting implementation and test cases.

In order to test the new herbstclient, the test suite features a
minimalist ipc server implementation which can be used to simulate wrong
behaviour and thus to stress the error handling in the client.