Use correct comparison function in 'complete' (#1289)
Thanks to the coverage reported, I've noticed that the function
parse_and_compare<unsigned long> was not used at all. The mistake was
introduced in 698dcd74ddbe96e96429c5ce874b2d099fd8c3cd.

It is hard to write a test case for it, but I've added some test case of
other yet non-covered parts of the error handling and completion of
'compare'.