Make 'rule' report errors during rule creation
When creating a new rule (via the `rule` or `apply_tmp_rule` command),
then a erroneous argument to a rule consequence (e.g. an invalid regular
expression) will be reported directly and will make the rule creation
fail. This helps finding mistakes earlier. Before, herbstluftwm printed
an error message during rule usage to its stderr, which is usually not
monitored by anybody.

On the negative side, this makes it now impossible to use 'toggle' in
boolean rule consequences. But I assume nobody uses this anyway, because
the result is of course highly unpredictable. Also, no rule consequence
can fail at run time at the moment, which produces dead code in
Rule::evaluate(), but I'm keeping this to be on the safe side.

As a positive side-effect, the redundancy in the applier implementation
in the Consequence class is reduced.

This change actually simplifies the test cases, because parse errors are
now reported directly.