Handle text properties more robustly (#1303)
This improves reading and writing of utf8 string list X11-properties,
and makes hlwm work even if the locale is partly broken, e.g. on some
fresh Manjaro installations (#1276).

When writing string lists in XConnection::setPropertyString(), we first
try the utf8-function, and if it fails, fall back to `XStdICCTextStyle`,
which seems to work even in the above mentioned scenario. Unfortunately,
I don't know how to reproduce this scenario in the CI, so this is not
covered code in the CI. (Here, we would need a test case for which
XSupportsLocale() in main.cpp fails).

Reading string lists is now centralized in
XConnection::getWindowPropertyTextList(). Formerly, we used
Xutf8TextPropertyToTextList() here, which unfortunately failed in the
setup of #1276. Interestingly, tools like `xprop` don't use it at all,
and instead call the plain `XGetWindowProperty()` and parse the string
list from there, and indeed this seems to work with or without working
locales (and it's tested thoroughly by the CI). For the sake of avoiding
redundancy, we can use this implementation in the IpcServer (it formerly
duplicated the source code) and when reading ordinary utf8 string
properties (XConnection::getWindowProperty()).