Client alias 'longest-minimized' and 'last-minimized'
This adds client aliases for accessing the longest and the most recently
minimized client, and thus provides a native way to un-minimize clients
with key bindings (either in a lifo or fifo way). Thus, this also works
when minimizing clients via the panel and thus is more universal than
restricting to the unminimize.sh script (which was written for the lifo
unminimize).

This solves another feature request raised in the discussion of #1014.