Update floating geometry if size hints change
If the size hints of a client change, then its floating geometry must be
updated. Also, this must happen even if the client is not floated at the
moment. To accomplish the latter, the applysizehints() function is
extended by an optional `force` parameter. The new test case verifies
both the floating and non-floating case.

This fixes #1322.