Set X11 property for client's floating state
The purpose of this change is to indicate a clients effective floating
state via X11 properties to make use of it in rules of X11 compositors.

The change happens in two steps: The first step is a new read-only
attribute `floating_effectively` which combines the client's single
window floating state with its tag's floating state. This attribute is
tested semantically. The second step is simply a signal connection that
sets or removes the X11 properties HLWM_FLOATING_WINDOW and
HLWM_TILING_WINDOW, depending on the value of `floating_effectively`.

This closes #1172.