Shorten window title to decoration width
Shorten a client's title such that it is at most as wide as the window
content, i.e. such that it does not reach into window border.

The new tests verify this by looking for the leftmost and rightmost
pixels belonging to the title. Unfortunately, proper utf8 characters
work in the test suite only with xft fonts. But in a manual xephyr
session, the plain X fonts also work with utf8 characters.