Update cursor shape and resize accordingly
When the mouse cursor is at a client's window border, the cursor shape
now changes to indicate the the direction in which the window resizes.
Also, I've updated the resize code itself to allow resizing where the
width or height is fixed and the client is resized only into one
dimension.

Unfortunately, the cursor preview isn't entirely correct for tiled
windows. Also does not seem to be possible to query the cursor shape in
test cases.

This closes #1138.