spawn: Print error message on exec failure (#1333)
This encapsulates the spawning of new processes in a utility function
that either returns the pid of the child process or a error message.
Thus, the 'spawn' command now prints a meaningful error message if the
exec failed.

Formerly, we manually closed the socket for the X11 connection on every
fork. Now, the X11 socket is marked as 'CLOEXEC' (close on exec), if it
hasn't been already by XOpenDisplay(). This eliminates another instance
of the global g_display variable.

Unfortunately, I don't know how to really test the low level error
messages for pipe() and fork(). But there are test cases for the other
error messages for 'spawn' and also a new test case for 'wmexec'.

In the long run, the spawnProcess() function should also be used in the
autostart invocation in the 'reload' command to also provide an error
message there.