tests: Close stdout/stderr after subprocess.Popen
When using subprocess.Popen() and piping stdout or stderr, the pipe
needs to be closed on shutdown.

For the present change, I've grep'ed for all uses of subprocess.Popen()
and added the necessary close() invocations. In one occasion, I could
simply replace subprocess.Popen() by subprocess.communicate() which
already takes care of this.

Unclosed file descriptors and similar issues are reported by passing
`-Wdefault` to pytest.

Thanks to @Flowdalic for reporting this issue.