Resize frames in both dimensions (#1348)
This implements a feature which was already indicated by the cursor shape:
resizing frames in both x and y direction. For this, it was necessary to
entirely change the drag handler for resizing frames.

As a little bonus, the special cursor shape persists during the drag.

The new test case essentially checks that resizing in both dimensions
works as expected. The existing testcase for 1-dimensional resize
could be simplified a lot using new attributes and the convenience
functions in the python bindings.