Fix doc for member-initialized attributes (#1384)
This fixes a wrongly nested if-statement in gendoc.py which caused the
doc string of Attributes to be hidden if they were member-initialized
(during the class definition). This affects the doc string of all theme
attributes.

The new test case checks that (and failed before the present change to
gendoc.py).