Extend 'foreach' by a --filter-name=REGEX flag (#1387)
This migrates the 'foreach' command to the ArgParse/CallOrComplete system which
makes it easy to add command line flags to the invocation syntax.
For this, the ArgParse class had to be extended slightly such that it
also support additional arguments after the positional arguments and
flags. In the 'foreach' example, those additional arguments specify the
nested command, i.e. the 'loop body'.