Extend foreach command by --unique and --recursive (#1394)
This adds two flags to foreach: --unique removes duplicates (or more
precisely, aliases) of objects and --recursive recurses into subobjects.
To ensure termination --recursive implicitly also uses --unique. The
recursion is done in a breadth-first way such that deeply nested objects
are visited later.

The main application for --unique is to avoid the duplicate
clients.focus and the main application of --recursive is to traverse all
frames in the tree layout.