Add attributes 'title_when' and 'title_depth'
This makes it configurable when to show the title bar. So far, there are
four options that essentially depend on the number of tabs.
So far, I've used padding_top to add space for characters in the title
that reach below the baseline. Since the space below the baseline should
be present only when the title is shown, I've added a new setting
title_depth.

This new feature was requested multiple times on the IRC and in the
discussion of #1097 and #1388.