Add settings.ellipsis
As requested in #1395, this adds an ellipsis, which is appended to the
window title if its to long for the available space. This requires that
we erase characters in the middle of a utf-8 string, and so we need to
take care of characters that amount to multiple bytes in the string.

This closes #1395.