Simplify urgency handling by using signals
This simplifies the handling of urgency handling and as a bonus, it
makes the urgency attribute writable.

The main idea behind this change is to separate the logic from the
communication with X11. One function (readWmHints) reads the urgency
hint from X11 to the attributes, and a signal propagates changes of the
attribute's value back to X11. The entire hlwm 'logic' only operates on
the urgency attribute (and so we can also grant the user write access to
it).

Also, this avoids a few X11 errors if a new client is already urgent.
Before this change, an urgent client triggered an early relayout, which
even led to the CI failure in
https://github.com/herbstluftwm/herbstluftwm/pull/1399.