Let active and inactive tabs have same text width (#1401)
If the title was centered or right-aligned, the title sometimes moved
slightly. This was caused by the text area having slightly different
widths for active and inactive tabs. The present change fixes this and
adds a test case for it.

Unfortunately, the test case is quite time consuming (8 seconds here),
because it takes 3 * 2 = 6 screenshots, but I think it's worth the
effort.