Move settings doc to in-source documentation (#1405)
This removes the section on settings from the man page and replaces it
with the automatically generated documentation on the attributes of the
settings object.

Unfortunately, the asciidoc syntax is a bit of a mess, and so it is very
complicated to correctly put longer doc strings containing code blocks
together. And in addition to that, escaping special characters is also a
mess, and so I'm now using html entities in the doc string of the
tree_style setting (for the time being, the help command just literally
prints the html entities, but the html and man page documentation
resolves them correctly).