Add theme options for unselected tabs (#1410)
In order to distinguish unselected from selected tabs, this adds new
attributes for controlling tab appearance. Those attributes can also be
empty, which will cause the normal attributes to be used (as before);
this makes the change 100% backwards compatible.

This implements what has been requested during the discussion of #1388
and closes #1389.