Do not map already visible unmanaged windows
In the MapNotify for an unmanaged window, manage_client() is called,
which should not call XMapWindow(), because the window is already
visible. When XMapWindow() was still called, this could trigger wrong
behaviour if a x11 client quickly maps and then unmaps the window,
because hlwm would then map the window again.

Concretely, this bugs triggered an endless loop when pressing Alt in
pcmanfm (one should only test this in Xephyr, otherwise the session
hangs).

The new test case quickly maps and unmaps a window, and verifies that
hlwm does not try to map the window again.