Fix input focus for preexisting clients
This correctly puts the input focus to clients that exist before hlwm.
This is a bug that was brought to the surface by PR #1403, because now,
hlwm only calls XSetInputFocus() if the focus actually changes.

The solution now is: the decoration window is 'unmapped' if and only if
the client window itself is 'unmapped'. This is true already for all
clients that show up during the hlwm exection, but wasn't true for
clients existing before. With this property, making the client visible
(in Client::set_visible()) will actually set the client to the 'mapped'
state, triggering a MapNotify event, whose handler then calls
XSetInputFocus() if necessary.