Clamp to XRectangle dimensions (#1423)
If inner_width, outer_width, and border_width are identical, then this
makes one of the rectangle heights negative. Since heights of XRectangle
are of type unsigned short, this caused a very large rectangle. To fix
this, the helper function toUnsigned() now correctly casts widths and
heights to unsigned short.

This fixes #1418