Drag floating clients between monitors (#1424)
This allows that floating clients can be dragged between monitors. When
moving the dragged client to another monitor, the monitors are redrawn
multiple times, and one flicker might occur, but I think one can live
with that.

While at it, I also removed another odd behaviour: if a tag with
floating windows is moved from a big to a small monitor, then the
windows do not 'glue' in the bottom right corner anymore, because the
floating position is now fixed when the drag starts.

Both new features have their test case.

This closes #1415 and #1245.