Preserve detected monitor order when deduplicating
The monitor detection must remove duplicates in the list of detected
outputs, because a monitor might be mirrored to another one. However,
removing duplicates must still preserve the order, especially if there
are no duplicates at all -- most importantly because the 'primary'
monitor is at index 0.

In the X serves used in the test suite (Xephyr, Xvfb), we unfortunately
can't specify the outputs' position, so we use that Xephyr places
outputs next to each other and Xvfb places them all at (0,0).

Also, this introduces a test case for randr (only one monitor works
here, but it's better than nothing).

This fixes #1363.