Use xmllint step of a2x for patching manpages (#1428)
This essentially reverts the cmake changes of f61211ad0f (#1396) and
instead injects the patch-manpage-xml.py via asciidoc's xmllint step.

The positive effect of this is that we do not need to know the path used
in ASCIIDOC_MANPAGE_XSL. Recently on arch linux, it was moved from

    /etc/asciidoc/docbook-xsl/manpage.xsl

to

    /usr/lib/python3.10/site-packages/asciidoc/resources/docbook-xsl/manpage.xsl

which of course depends on a lot of parameters (python version, etc).
Since I couldn't find a way to let asciidoc tell me the path, so having
the path a cmake parameter will presumably cause a lot of packaging
headache.

Instead, I noticed, that we could abuse the 'xmllint' step of the
asciidoc compilation to inject a tool that patches the man page xml.
As a side-effect, this changes make the documentation build again
successfully out of the box on arch linux.