Premultiply colors with alpha if compositing
This fixes the color handling of rgba colors. The compositing of X11
expects the true transparency colors to be 'premultiplied' with the
alpha value. That is, a color r=g=b=80 with 50% alpha should have
r=g=b=40 (see also [1,2]).

This bug was still present in colors for both frame decorations and
normal decorations.

Due to this premultiplication, colors would appear darker if the
premultiplication was applied also in absence of a composting manager
(aka compositor). Thus, this change also includes new code that detects
the presence and absence of a compositing manager at run time, using the
xfixes X11 extension library (which should be available on all system
that have X11).

The new test cases essentially checks that hlwm toggles this
premultiplication depending on whether a compisitor claims to exist.

On the implementation side, this color conversion was moved to
XConnection, and so a few Color member functions are not needed anymore.

This fixes #1382.

[1] https://stackoverflow.com/q/39505100/4400896
[2] https://stackoverflow.com/a/39927205/4400896