CI: wait for client acquiring input focus
This fixes a race condition in the test regarding keys_inactive: For
many applications, hlwm just tells the window that it should get the
input focus. Thus, we need to wait for this to finish before we can send
input events to the window.

While at it, I've removed the vacuous 'client_first' parameter in
test_keymask_for_existing_binds.