IPC: drop property events from queue early (#1467)
This is an attempt to fix #1448.

### What I think what the issue is ###
The ipc server listens for new client
request via property events and sends the reply back via properties
on the same window. However, the latter causes four further such
property events: one for the deletion of the request, and property
change events for stdout, stderr, status.

With local experiments for the extremely huge test case
`test_completable_commands`, I noticed that at the end of the test case,
these property events are still piled up in the event queue (sometimes
there are more than thousand of them). So the queue contained a long
list of event blocks of the shape:

    CreateNotify for winid=0x600001
    PropertyNotify, winid=0x600001, "_HERBST_IPC_ARGS", del=1
    PropertyNotify, winid=0x600001, "_HERBST_IPC_OUTPUT", del=0
    PropertyNotify, winid=0x600001, "_HERBST_IPC_ERROR", del=0
    PropertyNotify, winid=0x600001, "_HERBST_IPC_EXIT_STATUS", del=0
    DestroyNotify for winid=0x600001
    CreateNotify for winid=0x600001
    PropertyNotify, winid=0x600001, "_HERBST_IPC_ARGS", del=1
    PropertyNotify, winid=0x600001, "_HERBST_IPC_OUTPUT", del=0
    PropertyNotify, winid=0x600001, "_HERBST_IPC_ERROR", del=0
    PropertyNotify, winid=0x600001, "_HERBST_IPC_EXIT_STATUS", del=0
    DestroyNotify for winid=0x600001
    ...

Even more surprising, each such block mentioned the same window id,
because apparently, the window id was re-used by the next 'herbstclient'
because it seemed to be free again.

What I don't understand is, why hlwm processes a new client request
before consuming the self-triggered property. X11 seems to be aware of
this reordering, at least this would explain why each such block starts
with a CreateNotify and ends with a DestroyNotify: as if X11 wants to
emphasize that this 0x600001 is different from the 0x600001 mentioned
before.

This long queue (> 1000) then gets consumed eventually during test case
tear down, which explains the timeout that happened in late herbstclient
calls and when waiting for herbstluftwm to terminate.

### The workaround ###
Immediately after the ipc server sets or deletes the properties on the
client window, we remove all events triggered by it from the event
queue. I hope that XChangeProperty() places the event directly in the
local event queue without the round trip via the xserver (... otherwise
we would need some XSync() or so).

This change seems to make `test_completable_commands` reliable on my
local machine, and according to my debug output, the event queue is much
smaller (< 10) during test case teardown.