Add keybind for key release events (#1446)
This follows the suggestion of #1443 to implement release keybinds using
a fake modifier `Release`.

There are still some member variables that should be renamed. E.g. `grabbed`
should be `active`, but I didn't rename it because the diff is already quite
long.

Thanks to @madroach for suggesting things and for providing the first
implementation.