New setting 'show_frame_decorations'
This new settings gives a more fine-grain control when frame decorations
are shown. The new setting strictly generalizes the always_show_frame
setting, which now only exists for backwards-compatibility.

The feature was requested in #1183, where it was reported that many
people find the initial frame decorations of a single big frame ugly
when starting hlwm for the first time. To address this, it maybe makes
sense to hide it per default. Thus, the setting is set to
`focused_if_ambiguous` in the in the default autostart.