Add --bidirectional arg for keybind
If a keybind for a combo is added, the corresponding release bind is
removed automatically. So if e.g. some script binds 'Alt-x' and another
script binds 'Release-Alt-x', then not both keybindings are active.
Instead, the first Alt-x is removed when 'Release-Alt-x' is bound.

In the rare occassion where both press and release are bound, this new
'--bidirectional' needs to be used.