tests: Work around endianess in image data
The raw data of an image obtained from a screenshot depends on the
endianess. I couldn't make colormap.query_colors() work for depth 32 or
24, which would be the clean solution. Instead, we still parse the
binary image data manually and depend on the endianess of the system. In
principle, this could still be different from the image's endianess in
distributed X11 setups, but this should not happen when running the test
suite.

This fixes #1486.