IPC Server: fix type of command status
The Xlib function XChangeProperty() requires the data array to be
an array of 'long'. This causes wrong behaviour if long is larger than
int and if byte order is big-endian.

Unfortunately, I don't see how I can verify the now fixed issue in a
test case.

There is another issue in the status property: the atom should have type
XA_CARDINAL instead of XA_ATOM. Changing this would require the
corresponding change in herbstclient. Strictly speaking, this would be
an incompatible change in the IPC protocoll.

This fixes #1414.