Make ewmhnotify behave as expected
The purpose of disabling `ewmhnotify` is to set the hlwm-fullscreen
state independent of the ewmh fullscreen state (either to have one
enabled and the other disabled, or vice versa). To make this more easy
to use, this de-coupling should happen in both directions.

The main purpose seems to make firefox think that it is in fullscreen
state (such that it hides the gui) without firefix covering the whole
monitor (#1488).