IPC Server: fix type of command status (#1502)
The Xlib function XChangeProperty() requires the data array to be
an array of 'long'. This causes wrong behaviour if long is larger than
int and if byte order is big-endian.

I could reproduce the issue #1414 on the ppc64 system gcc203 of the
gcc compile farm. Unfortunately, I don't see how I can verify the now
fixed issue in a test case.

There is another issue in the status property: the atom should have type
XA_CARDINAL instead of XA_ATOM. Changing this would require the
corresponding change in herbstclient. Strictly speaking, this would be
an incompatible change in the IPC protocol.

This fixes #1414.