Improve client css class names and debugging (#1516)
For easier CSS theme debugging, this prints the widget tree of a given
client decoration. Since this exposes the used class names to the user,
this change also revises some of the names and ensures that each widget
has at least one class name. Also, we can use this command for examples
in the theme documentation in the future.