Make geometry based panel detection an attribute (#1521)
If a panel does not define _NET_WM_STRUT, we formerly just used it's window size for determining the reserved space at the screen edge. This caused issues such as #1475. By making it a setting, only those users who really need it can activate it.

While at it, I've fixed a bug in `Signal::connect()` arising from a wrong use of `std::bind()`.

This closes #1475 and #1175.