Re-enable using double hyphens before rule options
Previously, double hyphens were stripped in the tokenizeArg function
which is only called after checking if the argument is "prepend",
"once", "printlabel", "not", "!", or "fixedwidth". This failed to match on
"--once", "--prepend", etc..., so the argument is eventually passed to
tokenizeArg. But tokenizeArg expects all of its arguments to contain a = or ~
(everything else was supposed to be handled already). This resulted in
an error.

Now, the stripping of hyphens is done at the beginning of rule parsing,
and after manually moving the loop pointer for "not" and "!".

This closes #1487