Stick to old python-xlib in CI until bugfix
In python-xlib 0.32, the bug https://github.com/python-xlib/python-xlib/pull/242
was introduced which caused `test_floatplacement_smart_create_many` to
fail deterministically. There already is a pull request in `python-xlib`
with a fix, but it is not merged yet. So for the time being, let's
simply use the old python-xlib version.