Add sticky window/client support (#1542)
This PR implements sticky windows/clients as discussed in #976.

Anything I use this feature for works without issues, but I'll test a few days/weeks for bugs.

Clients always stay on the monitor they are on. This is how bspwm handles sticky windows too. When tags are swapped between two monitors, the sticky windows are swapped too, so that they stay on the correct monitor. Sticky windows stay focused when switching tags, but do not gain focus if they were previously not foucsed. Rule support is added too and works.